Surface sensor-staleness threshold on the Thermostats page
The "Sensor-staleness threshold" card added in #224 lived in
pages/Settings.tsx, which was never wired into the router or the
Nav — making both that card and the orphan page's per-thermostat
duplicate editor unreachable from the UI. Move the card onto the
Thermostats page (above the per-thermostat list, next to the other
system-wide cards), and delete the orphan page and its tests; the
per-thermostat fields are already covered by the canonical editor
in Thermostats.tsx.
